Fix most problems, remove unused code and remove SDK header
Add missing CoInitialize and CoUninitialize calls.
Change to INPVector2 to set custom distance between reflectors to try to support TrackClip PRO
Remove optitrack.h and replace by an include directory setting in the project that point to the include folder in the OptiTrack SDK.
Do not use the optitrack.h file from the previous commit. It was from another SDK version so it is not compatible with the OptiTrack.dll file for this version.
Update readme and point to new repository.
Add euler to quaternion conversion (does not work well).
Add error output to initialization function.